dev_requirements.txt
nose==1.1.2
django-nose==0.1.3
pyyaml==3.10
